Add some color to the page; try to add a box around each span
Note that when you try to add a box around the spans, you'll end up with elements that have no width. This is because inline elements don't have an
internal width. In order to style the cells, you'll need to change its `display` attribute value.

Use the chrome inspector to explore what some of the values of `display` do to an element. In particular, explore `inline`, `block`, and
`inline-block`. Choose the value that seems to fit best (probably `inline-bock`).