Introduce positioning by moving the character to a specific position
Use position absolute, and then open up the Chrome inspector again to see what happens as you change the values of left and top. Also notice that when we say
the top is 0, the character goes all the way to the top of the window, and when the left is 0, the character goes all the way to the left of the window.