Change the grid to position relative
Give the grid element a class and change its positioning type to `relative`. Explain that this makes any elements that are inside of the grid element
and positioned with `absolute` use this element as the reference. Now, since this change, a top of 0 make the character go all the way to the top of the
_grid_, and left of 0 makes the character go all the way to the left of the grid. Explore this with the Chrome inspector. How do you make the character go
_above_ the grid?